:sparkles: Support ContextLoader config in @⁠JUnit[Web]Config
Prior to this commit, @⁠JUnitConfig and @⁠JUnitWebConfig
did not declare `loader` attributes that alias @⁠ContextConfiguration's
`loader` attribute. Consequently, it was not possible to configure a
custom ContextLoader via those annotations.

The lack of those `loader` attributes was an oversight, and this commit
introduces them to support custom ContextLoader configuration directly
via the @⁠JUnitConfig and @⁠JUnitWebConfig annotations.